requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
autopep8==2.0.2
billiard==3.6.4.0
black==22.12.0
boto3==1.24.26
botocore==1.27.96
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configparser==5.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-prometheus==2.2.0
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gprof2dot==2022.7.29
grpcio==1.51.1
grpcio-tools==1.51.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
pika==1.3.1
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sentry-sdk==1.19.1
six==1.16.0
django-silk
sniffio==1.3.0
sqlparse==0.4.3
structlog==22.3.0
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
